requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-ldap
PyJWT
python-multipart
pandas
python-dotenv